Fixed Linting Errors
