+++ algos/common/agents.py
@@ -27,7 +27,7 @@ class Agent:
 
 
 class ValueAgent(Agent):
-
+    """Value based agent that returns an action based on the Q values from the network"""
     def __init__(self, net: nn.Module, action_space: int, eps_start: float = 1.0,
                  eps_end: float = 0.2, eps_frames: float = 1000):
         super().__init__(net)
